TST: simplify (and fix) field ambiguity testing
