+++ yt/fields/tests/test_ambiguous_fields.py
@@ -1,6 +1,3 @@
-import warnings
-from collections import defaultdict
-
 import pytest
 
 from yt._maintenance.deprecation import VisibleDeprecationWarning
@@ -11,34 +8,29 @@ def test_ambiguous_fails():
     ds = fake_amr_ds(particles=10)
     msg = "The requested field name '{}' is ambiguous"
 
-    fnames_to_ftype = defaultdict(list)
-    for ftype, fname in ds.field_list:
-        fnames_to_ftype[fname].append(ftype)
+    def _mock_field(field, data):
+        return data["ones"]
 
-    ambiguous_fnames = [
-        fname for fname, ftypes in fnames_to_ftype.items() if len(ftypes) > 1
-    ]
-    unambiguous_fnames = [
-        fname for fname, ftypes in fnames_to_ftype.items() if len(ftypes) == 1
-    ] + ["zeros", "ones"]
-
-    # Make sure we are testing both cases with a few cases
-    assert len(ambiguous_fnames) > 2
-    assert len(unambiguous_fnames) > 2
+    # create a pair of ambiguous field
+    ds.add_field(("io", "mock_field"), _mock_field, sampling_type="cell")
+    ds.add_field(("gas", "mock_field"), _mock_field, sampling_type="cell")
 
     # Test warnings are issued for ambiguous fields
-    for fname in ambiguous_fnames:
-        with pytest.warns(VisibleDeprecationWarning, match=msg.format(fname)):
-            ds.r[fname]
+    with pytest.warns(VisibleDeprecationWarning, match=msg.format("mock_field")):
+        ds.r["mock_field"]
 
-    # Test no warnings are issued for single fname access that aren't ambiguous
-    for fname in unambiguous_fnames:
-        with warnings.catch_warnings():
-            warnings.simplefilter("error")
-            ds.r[fname]
 
-    # Test no warning are issued for tuple access
-    for ftype, fname in ds.field_list:
-        with warnings.catch_warnings():
-            warnings.simplefilter("error")
-            ds.r[(ftype, fname)]
+def test_nameonly_field_with_all_aliases_candidates():
+    # see https://github.com/yt-project/yt/issues/3839
+    ds = fake_amr_ds(fields=["density"], units=["g/cm**3"])
+
+    # here we rely on implementations details from fake_amr_ds,
+    # so we verify that it provides the appropriate conditions
+    # for the actual test.
+    candidates = [f for f in ds.derived_field_list if f[1] == "density"]
+    assert len(candidates) == 2
+    fi = ds.field_info
+    assert fi[candidates[0]].is_alias_to(fi[candidates[1]])
+
+    # this is the actual test (check that no error or warning is raised)
+    ds.all_data()["density"]
